# docs/operations/iam-api.openapi.yaml
openapi: 3.1.0
info:
  title: IMS IAM API
  version: 1.0.0
  description: OpenAPI 3.1 contract for the IMS Identity & Access Management `/api/v1` surface.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204': { description: Logout success }
        '401': { $ref: '#/components/responses/Problem401' }
...
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
...
  /permissions:
    get:
      tags: [permissions]
      summary: List permissions
      responses:
        '200':
          description: Permission list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPermissionsEnvelope'
    post:
      tags: [permissions]
      summary: Create permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          description: Permission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionEnvelope'
...
  /reports/iam/{reportType}/export:
    post:
      tags: [reports]
      summary: Create report export job
      parameters:
        - $ref: '#/components/parameters/reportType'
      responses:
        '202':
          description: Export job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobEnvelope'
...
  /dashboards/iam/{dashboardType}:
    get:
      tags: [dashboards]
      summary: Get IAM dashboard
      parameters:
        - $ref: '#/components/parameters/dashboardType'
      responses:
        '200':
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    reportType:
      name: reportType
      in: path
      required: true
      schema: { type: string }
    dashboardType:
      name: dashboardType
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem400:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem401:
      description: Authentication error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, errorCode]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errorCode: { type: string }
        correlationId: { type: string, nullable: true }
...
    LogoutRequest:
      type: object
      properties:
        sessionId: { type: string }
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
        confirmPassword: { type: string }
...
    CreatePermissionRequest:
      type: object
      additionalProperties: true
...
    PermissionEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            permission: { type: object, additionalProperties: true }
    PaginatedPermissionsEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items: { type: object, additionalProperties: true }
            total: { type: integer }
...
    ExportJobEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            job: { type: object, additionalProperties: true }
...
    DashboardEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true